+++ gcsio/src/main/java/com/google/cloud/hadoop/gcsio/PrefixMappedItemCache.java
@@ -119,6 +119,17 @@ public class PrefixMappedItemCache {
   }
 
   /**
+   * Checks where items with specified prefix are in the cache.
+   *
+   * @param id the prefix resource id of the cached items to check.
+   * @return true if items with provided prefix are cached, false otherwise.
+   */
+  public synchronized boolean isPrefixCached(StorageResourceId id) {
+    PrefixKey key = new PrefixKey(id.getBucketName(), id.getObjectName());
+    return !getPrefixSubMap(itemMap, key).isEmpty();
+  }
+
+  /**
    * Invalidates all cached items and lists associated with the given bucket.
    *
    * @param bucket the bucket to invalidate. This must not be null.